move soft_neq to internals for now
