+++ chirho/explainable/internals/defaults.py
@@ -2,7 +2,12 @@ import functools
 from typing import TypeVar
 
 import pyro
+import pyro.distributions as dist
+import pyro.distributions.constraints as constraints
 import torch
+from torch.distributions import biject_to
+
+from chirho.indexed.ops import cond
 
 S = TypeVar("S")
 T = TypeVar("T")
@@ -60,3 +65,118 @@ def _uniform_proposal_integer(
         )
     n = support.upper_bound - support.lower_bound + 1
     return pyro.distributions.Categorical(probs=torch.ones((n,)))
+
+
+@functools.singledispatch
+def soft_eq(support: constraints.Constraint, v1: T, v2: T, **kwargs) -> torch.Tensor:
+    """
+    Computes soft equality between two values `v1` and `v2` given a distribution constraint `support`.
+    Returns a negative value if there is a difference (the larger the difference, the lower the value)
+    and tends to `Norm(0,1).log_prob(0)` as `v1` and `v2` tend to each other,
+    except for when the support is boolean, in which case it returns `0.0` if the values are equal
+    and a large negative number (`eps`) otherwise.
+
+    :param support: distribution constraint (`real`/`boolean`/`positive`/`interval`).
+    :params v1, v2: the values to be compared.
+    :param kwargs: Additional keywords arguments:
+        - for boolean, the function expects `eps` to set a large negative value for.
+        - For interval and real constraints, `scale` adjusts the softness of the inequality.
+    :return: A tensor of log probabilities capturing the soft equality between `v1` and `v2`.
+    :raises TypeError: If boolean tensors have different data types.
+    """
+    if not isinstance(v1, torch.Tensor) or not isinstance(v2, torch.Tensor):
+        raise NotImplementedError("Soft equality is only implemented for tensors.")
+    elif support.is_discrete:
+        raise NotImplementedError(
+            "Soft equality is not implemented for arbitrary discrete distributions."
+        )
+    elif support is constraints.real:  # base case
+        scale = kwargs.get("scale", 0.1)
+        return dist.Normal(0.0, scale).log_prob(v1 - v2)
+    else:
+        tfm = biject_to(support)
+        v1_inv = tfm.inv(v1)
+        ldj = tfm.log_abs_det_jacobian(v1_inv, v1)
+        v2_inv = tfm.inv(v2)
+        ldj = ldj + tfm.log_abs_det_jacobian(v2_inv, v2)
+        for _ in range(tfm.codomain.event_dim - tfm.domain.event_dim):
+            ldj = torch.sum(ldj, dim=-1)
+        return soft_eq(tfm.domain, v1_inv, v2_inv, **kwargs) + ldj
+
+
+@soft_eq.register
+def _soft_eq_independent(support: constraints.independent, v1: T, v2: T, **kwargs):
+    result = soft_eq(support.base_constraint, v1, v2, **kwargs)
+    for _ in range(support.reinterpreted_batch_ndims):
+        result = torch.sum(result, dim=-1)
+    return result
+
+
+@soft_eq.register(type(constraints.boolean))
+def _soft_eq_boolean(
+    support: constraints.Constraint, v1: torch.Tensor, v2: torch.Tensor, **kwargs
+):
+    assert support is constraints.boolean
+    scale = kwargs.get("scale", 0.1)
+    return torch.log(cond(scale, 1 - scale, v1 == v2, event_dim=0))
+
+
+@soft_eq.register
+def _soft_eq_integer_interval(
+    support: constraints.integer_interval, v1: torch.Tensor, v2: torch.Tensor, **kwargs
+):
+    scale = kwargs.get("scale", 0.1)
+    width = support.upper_bound - support.lower_bound + 1
+    return dist.Binomial(total_count=width, probs=scale).log_prob(torch.abs(v1 - v2))
+
+
+@soft_eq.register(type(constraints.integer))
+def _soft_eq_integer(
+    support: constraints.Constraint, v1: torch.Tensor, v2: torch.Tensor, **kwargs
+):
+    scale = kwargs.get("scale", 0.1)
+    return dist.Poisson(rate=scale).log_prob(torch.abs(v1 - v2))
+
+
+@soft_eq.register(type(constraints.positive_integer))
+@soft_eq.register(type(constraints.nonnegative_integer))
+def _soft_eq_positive_integer(support: constraints.Constraint, v1: T, v2: T, **kwargs):
+    return soft_eq(constraints.integer, v1, v2, **kwargs)
+
+
+@functools.singledispatch
+def soft_neq(support: constraints.Constraint, v1: T, v2: T, **kwargs) -> torch.Tensor:
+    """
+    Computes soft inequality between two values `v1` and `v2` given a distribution constraint `support`.
+    Tends to zero as the difference between the value increases, and tends to
+    `-eps / (Norm(0,scale).log_prob(0) - 1e-10)` as `v1` and `v2` tend to each other.
+
+    :param support: distribution constraint (`real`/`boolean`/`positive`/`interval`).
+    :params v1, v2: the values to be compared.
+    :param kwargs: Additional keywords arguments:
+        - for boolean, the function expects `eps` to set a large negative value for.
+        - For interval and real constraints, the function expects `eps` to fix the minimal value and
+        `scale` to adjust the softness of the inequality.
+    :return: A tensor of log probabilities capturing the soft inequality between `v1` and `v2`.
+    :raises TypeError: If boolean tensors have different data types.
+    :raises ValueError: If the specified scale is less than `1 / sqrt(2 * pi)`, to ensure that the log
+                        probabilities used in calculations are are nonpositive.
+    """
+    if not isinstance(v1, torch.Tensor) or not isinstance(v2, torch.Tensor):
+        raise NotImplementedError("Soft equality is only implemented for tensors.")
+    elif support.is_discrete:  # for discrete pmf, soft_neq = 1 - soft_eq (in log space)
+        return torch.log(-torch.expm1(soft_eq(support, v1, v2, **kwargs)))
+    elif support is constraints.real:  # base case
+        scale = kwargs.get("scale", 0.1)
+        return torch.log(2 * dist.Normal(0.0, scale).cdf(torch.abs(v1 - v2)) - 1)
+    else:
+        tfm = biject_to(support)
+        return soft_neq(tfm.domain, tfm.inv(v1), tfm.inv(v2), **kwargs)
+
+
+@soft_neq.register
+def _soft_neq_independent(support: constraints.independent, v1: T, v2: T, **kwargs):
+    result = soft_neq(support.base_constraint, v1, v2, **kwargs)
+    for _ in range(support.reinterpreted_batch_ndims):
+        result = torch.sum(result, dim=-1)
+    return result
